requirements.txt
capybara-docsaid
chameleon-docsaid
fire
ipython
lightning
albumentations